requirements.txt
appdirs
pandas
PyYAML
requests
setuptools
pre-commit
